Energy-trader-main/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
web3>=6.11.0
pycryptodome>=3.20.0
requests>=2.31.0
pydantic>=2.5.0
python-dotenv>=1.0.0
